+++ src/main/java/io/kontur/eventapi/nifc/episodecomposition/NifcEpisodeCombinator.java
@@ -14,7 +14,9 @@ import java.util.Set;
 
 import static io.kontur.eventapi.nifc.converter.NifcDataLakeConverter.NIFC_LOCATIONS_PROVIDER;
 import static io.kontur.eventapi.nifc.converter.NifcDataLakeConverter.NIFC_PERIMETERS_PROVIDER;
+import static io.kontur.eventapi.util.GeometryUtil.isEqualGeometries;
 import static java.util.Collections.emptyList;
+import static org.apache.commons.lang3.StringUtils.equalsIgnoreCase;
 
 @Component
 public class NifcEpisodeCombinator extends WildfireEpisodeCombinator {
@@ -37,6 +39,10 @@ public class NifcEpisodeCombinator extends WildfireEpisodeCombinator {
         episode.setObservations(mapObservationsToIDs(episodeObservations));
         episode.setGeometries(computeEpisodeGeometries(episodeObservations));
         episode.setDescription(latestObservation.getDescription());
+        List<FeedEpisode> existing = feedData.getEpisodes();
+        if (!existing.isEmpty() && sameEpisode(existing.get(existing.size() - 1), episode)) {
+            return emptyList();
+        }
         return List.of(episode);
     }
 
@@ -47,6 +53,13 @@ public class NifcEpisodeCombinator extends WildfireEpisodeCombinator {
                 .distinct()
                 .toArray(Feature[]::new);
         return new FeatureCollection(features);
+    }
+
+    private boolean sameEpisode(FeedEpisode ep1, FeedEpisode ep2) {
+        return equalsIgnoreCase(ep1.getName(), ep2.getName())
+                && ep1.getSeverity() == ep2.getSeverity()
+                && equalsIgnoreCase(ep1.getLocation(), ep2.getLocation())
+                && isEqualGeometries(ep1.getGeometries(), ep2.getGeometries());
     }
 
 }